notebooks/requirements.txt
ultralytics
opencv-python
pyyaml